test(#3515): split the wait-census coverage out of budgets.rs (#1135)
budgets.rs was 40 lines from the 1500-line test threshold, and round 16 has to
add an assert and a test to it. Split by SUBJECT: everything about
assert_census_matches_run moves to budgets/census_tests.rs, the one deadline's
arithmetic stays. No test is added or removed here (35 before and after).
